Dangerous code checked in.  Pointer instances can point to a single data item, or they can point to an array of values.
Back in the distant past ;-), I had a SetSize() method, which would
enable indexes != 0. Don't know when and why I removed it, but I need
it again.  Should it have to be enabled explicitely?  Would be better,
it seems.

Note that there's no corresponding ass_item implementation with
indexes != 0. This means, we can get and change the items the pointer
points to, if they are mutable, but not replace the items by different
ones.

--HG--
branch : ctypes
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/ctypes/trunk/ctypes%4047381