Define a new protocol for function's restype.  If it is a ctypes' type, it is handled in the usual way.  But if it additionally has an _check_retval_ attribute, this is called with the result, and the return value of this call is the functions return value.
This allows to define HRESULT as a ctypes' type, because it can check
for failure in it's _check_retval_ static method, and raise a
WindowsError.

Although HRESULT is defined in python code (since it is derived from
c_long), the _check_retval_ method is implemented in C - in this way
the traceback won't include the _check_retval_ method definition itself.

--HG--
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/ctypes/trunk%4048697