* ctypes/wrap/codegenerator.py: sort the generated definitions according to location in the source code.
* ctypes/wrap/codegenerator.py: find_dllname() and get_sharedlib()
is not needed anymore (replaced by the changes in xml2py.py). In
the generated code load dll's explicitly based on name and
version of the dll.

* ctypes/wrap/xml2py.py: added classes to load a list of libraries
(given by the option -l), to find the library in which a function
is defined and to supply the information needed to load that
library in the generated code.

--HG--
branch : ctypes
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/ctypes/trunk/ctypes%4049349